90% feito so falta o editar

+++ index.js
@@ -1,5 +1,33 @@
-function excluir (id) {
-    fetch('http://localhost:8000/compras'+id, {
+function Inserir(){
+    event.preventDefault();
+
+    let dados = {
+        "item": input_item.value,
+        "quantidade": parseInt(input_quantidade.value)
+    };
+
+    fetch('http://localhost:8000/compras', {
+        method: 'POST',
+        body: JSON.stringify(dados),
+        headers: {
+            'Content-Type': 'application/Json'
+        }
+    })
+    .then(resposta => resposta.json())
+    .then(resposta => atualizarlista());
+
+    form_add.reset();
+}
+
+
+async function excluir (id) {
+    let resposta = confirm('Vc tem certeza?')
+    
+    if (resposta !== true) {
+        return;
+    }
+
+    fetch('http://localhost:8000/compras/'+id, {
         method: 'DELETE'
     });
 
@@ -8,6 +36,7 @@ function excluir (id) {
 
 
 function atualizarlista() {
+    tabela_compras.innerHTML = '';
     fetch('http://localhost:8000/compras')
     .then(function (resposta) {
         return resposta.json();
@@ -20,6 +49,10 @@ function atualizarlista() {
                 <td>${cadaItem.item}</td>
                 <td>${cadaItem.quantidade}</td>
                 <td>
+                    <button class="btn btn-warning">
+                        Editar
+                    </button>
+
                     <button onclick="excluir(${cadaItem.id})" class="btn btn-danger">
                         Excluir
                     </button>
